relocate & reorganize
